Refactor Orchestrator and Consumer and OrchestratorBackend and ConsumerBackend so the Orchestrator and Consumer can live in different places
