+++ mindtrace/jobs/mindtrace/jobs/local/consumer_backend.py
@@ -1,24 +1,23 @@
 import time
-from queue import Empty
-from typing import Optional, Callable
+from typing import TYPE_CHECKING
 
 from mindtrace.jobs.base.consumer_base import ConsumerBackendBase
 from mindtrace.jobs.utils.checks import ifnone
 
+if TYPE_CHECKING:
+    from mindtrace.jobs.local.client import LocalClient
+
 class LocalConsumerBackend(ConsumerBackendBase):
     """Local in-memory consumer backend."""
     
-    def __init__(self, queue_name: str, orchestrator, run_method: Optional[Callable] = None, 
-                 poll_timeout: float = 1):
-        super().__init__(queue_name, orchestrator, run_method)
+    def __init__(self, queue_name: str, consumer_frontend, orchestrator: "LocalClient", poll_timeout: float = 1):
+        super().__init__(queue_name, consumer_frontend)
         self.poll_timeout = poll_timeout
+        self.orchestrator = orchestrator
         self.queues = [queue_name] if queue_name else []
     
     def consume(self, num_messages: int = 0, *, queues: str | list[str] | None = None, block: bool = True, **kwargs) -> None:
         """Consume messages from the local queue(s)."""
-        if not self.run_method:
-            raise RuntimeError("No run method set.")
-        
         if isinstance(queues, str):
             queues = [queues]
         queues = ifnone(queues, default=self.queues)
@@ -65,7 +64,7 @@ class LocalConsumerBackend(ConsumerBackendBase):
         """Process a single message."""
         if isinstance(message, dict):
             try:
-                self.run_method(message)
+                self.consumer_frontend.run(message)
                 job_id = message.get('id', 'unknown')
                 self.logger.debug(f"Successfully processed dict job {job_id}")
                 return True
