+++ mindtrace/jobs/mindtrace/jobs/redis/consumer_backend.py
@@ -1,23 +1,24 @@
+import json
 import time
-from typing import Optional, Callable
+from queue import Empty
+from typing import Optional
 
 from mindtrace.jobs.base.consumer_base import ConsumerBackendBase
+from mindtrace.jobs.redis.connection import RedisConnection
 from mindtrace.jobs.utils.checks import ifnone
+
 
 class RedisConsumerBackend(ConsumerBackendBase):
     """Redis consumer backend with blocking operations."""
     
-    def __init__(self, queue_name: str, orchestrator, run_method: Optional[Callable] = None, 
-                 poll_timeout: int = 5):
-        super().__init__(queue_name, orchestrator, run_method)
+    def __init__(self, queue_name: str, consumer_frontend, host: str, port: int, db: int, poll_timeout: int = 5):
+        super().__init__(queue_name, consumer_frontend)
         self.poll_timeout = poll_timeout
         self.queues = [queue_name] if queue_name else []
-    
+        self.connection = RedisConnection(host=host, port=port, db=db)
+
     def consume(self, num_messages: int = 0, *, queues: str | list[str] | None = None, block: bool = True, **kwargs) -> None:
         """Consume messages from Redis queue(s)."""
-        if not self.run_method:
-            raise RuntimeError("No run method set.")
-        
         if isinstance(queues, str):
             queues = [queues]
         queues = ifnone(queues, default=self.queues)
@@ -27,7 +28,7 @@ class RedisConsumerBackend(ConsumerBackendBase):
             while num_messages == 0 or messages_consumed < num_messages:
                 for queue in queues:
                     try:
-                        message = self.orchestrator.receive_message(queue, block=block, timeout=self.poll_timeout)
+                        message = self.receive_message(queue, block=block, timeout=self.poll_timeout)
                         if message:
                             self.logger.debug(f"Received message from queue '{queue}': processing {messages_consumed + 1}")
                             if self.process_message(message):
@@ -48,7 +49,7 @@ class RedisConsumerBackend(ConsumerBackendBase):
         """Process a single message."""
         if isinstance(message, dict):
             try:
-                self.run_method(message)
+                self.consumer_frontend.run(message)
                 job_id = message.get('id', 'unknown')
                 self.logger.debug(f"Successfully processed dict job {job_id}")
                 return True
@@ -67,7 +68,7 @@ class RedisConsumerBackend(ConsumerBackendBase):
             queues = [queues]
         queues = ifnone(queues, default=self.queues)
         
-        while any(self.orchestrator.count_queue_messages(q) > 0 for q in queues):
+        while any(self.connection.count_queue_messages(q) > 0 for q in queues):
             self.consume(num_messages=1, queues=queues, block=block)
             
         self.logger.info(f"Stopped consuming messages from queues: {queues} (queues empty).")
@@ -75,3 +76,27 @@ class RedisConsumerBackend(ConsumerBackendBase):
     def set_poll_timeout(self, timeout: int) -> None:
         """Set the polling timeout for Redis operations."""
         self.poll_timeout = timeout 
+
+    
+    def receive_message(self, queue_name: str, **kwargs) -> Optional[dict]:
+        """Retrieve a message from a specified Redis queue.
+
+        Returns the message as a dict.
+        """
+        with self.connection._local_lock:
+            if queue_name not in self.connection.queues:
+                raise KeyError(f"Queue '{queue_name}' is not declared.")
+            instance = self.connection.queues[queue_name]
+        try:
+            if hasattr(instance, "get"):
+                raw_message = instance.get(block=False, timeout=None)
+            elif hasattr(instance, "pop"):
+                raw_message = instance.pop(block=False, timeout=None)
+            else:
+                raise Exception("Queue type does not support receiving messages.")
+            message_dict = json.loads(raw_message)
+            return message_dict
+        except Empty:
+            return None
+        except Exception:
+            return None
